Rename NftsTableV2 -> NftsTable.
