+++ src/components/account/NftsTable.vue
